+++ apps/laboratory/tests/shared/fixtures/w3m-smart-account-fixture.ts
@@ -1,16 +1,19 @@
-import { test as base } from '@playwright/test'
 import type { ModalFixture } from './w3m-fixture'
 import { Email } from '../utils/email'
 import { ModalWalletPage } from '../pages/ModalWalletPage'
 import { ModalWalletValidator } from '../validators/ModalWalletValidator'
 import type { ModalPage } from '../pages/ModalPage'
+import { timingFixture } from './timing-fixture'
 
 // Test Modal + Smart Account
-export const testModalSmartAccount = base.extend<ModalFixture & { slowModalPage: ModalPage }>({
+export const testModalSmartAccount = timingFixture.extend<
+  ModalFixture & { slowModalPage: ModalPage }
+>({
   library: ['wagmi', { option: true }],
   modalPage: [
     async ({ page, library, context }, use, testInfo) => {
       const modalPage = new ModalWalletPage(page, library)
+      const modalValidator = new ModalWalletValidator(page)
       await modalPage.load()
 
       const mailsacApiKey = process.env['MAILSAC_API_KEY']
@@ -24,6 +27,7 @@ export const testModalSmartAccount = base.extend<ModalFixture & { slowModalPage:
       await modalPage.openAccount()
       await modalPage.openSettings()
       await modalPage.switchNetwork('Sepolia')
+      await modalValidator.expectSwitchedNetwork('Sepolia')
       await modalPage.closeModal()
       await use(modalPage)
     },