+++ packages/core/tests/controllers/SendController.test.ts
@@ -23,7 +23,7 @@ const receiverProfileImageUrl = 'https://ipfs.com/0x123.png'
 // -- Tests --------------------------------------------------------------------
 describe('SendController', () => {
   it('should have valid default state', () => {
-    expect(SendController.state).toEqual({})
+    expect(SendController.state).toEqual({ loading: false })
   })
 
   it('should update state correctly on setToken()', () => {
@@ -53,6 +53,6 @@ describe('SendController', () => {
 
   it('should update state correctly on resetSend()', () => {
     SendController.resetSend()
-    expect(SendController.state).toEqual({})
+    expect(SendController.state).toEqual({ loading: false })
   })
 })