+++ packages/ethers/src/client.ts
@@ -10,12 +10,13 @@ import type {
   NetworkControllerClient,
   PublicStateControllerState,
   SendTransactionArgs,
-  Token
+  Token,
+  WriteContractArgs
 } from '@web3modal/scaffold'
 import { Web3ModalScaffold } from '@web3modal/scaffold'
 import { ConstantsUtil, PresetsUtil, HelpersUtil } from '@web3modal/scaffold-utils'
 import { AccountController, NetworkController } from '@web3modal/core'
-import EthereumProvider from '@walletconnect/ethereum-provider'
+import EthereumProvider, { OPTIONAL_METHODS } from '@walletconnect/ethereum-provider'
 import type { Web3ModalSIWEClient } from '@web3modal/siwe'
 import type {
   Address,
@@ -31,7 +32,10 @@ import {
   InfuraProvider,
   getAddress as getOriginalAddress,
   parseUnits,
-  formatUnits
+  formatUnits,
+  JsonRpcSigner,
+  BrowserProvider,
+  Contract
 } from 'ethers'
 import {
   EthersConstantsUtil,
@@ -47,11 +51,8 @@ import {
   W3mFrameConstants
 } from '@web3modal/wallet'
 import type { CombinedProvider } from '@web3modal/scaffold-utils/ethers'
-import { BrowserProvider } from 'ethers'
-import { JsonRpcSigner } from 'ethers'
 import { NetworkUtil } from '@web3modal/common'
 import type { W3mFrameTypes } from '@web3modal/wallet'
-
 // -- Types ---------------------------------------------------------------------
 export interface Web3ModalClientOptions extends Omit<LibraryOptions, 'defaultChain' | 'tokens'> {
   ethersConfig: ProviderType
@@ -189,7 +190,51 @@ export class Web3Modal extends Web3ModalScaffold {
           onUri(uri)
         })
 
-        await WalletConnectProvider.connect()
+        if (siweConfig?.options?.enabled) {
+          const { SIWEController, getDidChainId, getDidAddress } = await import('@web3modal/siwe')
+          const result = await WalletConnectProvider.authenticate({
+            nonce: await siweConfig.getNonce(),
+            methods: OPTIONAL_METHODS,
+            ...(await siweConfig.getMessageParams())
+          })
+          // Auths is an array of signed CACAO objects https://github.com/ChainAgnostic/CAIPs/blob/main/CAIPs/caip-74.md
+          const signedCacao = result?.auths?.[0]
+          if (signedCacao) {
+            const { p, s } = signedCacao
+            const chainId = getDidChainId(p.iss)
+            const address = getDidAddress(p.iss)
+            if (address && chainId) {
+              SIWEController.setSession({
+                address,
+                chainId: parseInt(chainId, 10)
+              })
+            }
+            try {
+              // Kicks off verifyMessage and populates external states
+              const message = WalletConnectProvider.signer.client.formatAuthMessage({
+                request: p,
+                iss: p.iss
+              })
+
+              await SIWEController.verifyMessage({
+                message,
+                signature: s.s,
+                cacao: signedCacao
+              })
+            } catch (error) {
+              // eslint-disable-next-line no-console
+              console.error('Error verifying message', error)
+              // eslint-disable-next-line no-console
+              await WalletConnectProvider.disconnect().catch(console.error)
+              // eslint-disable-next-line no-console
+              await SIWEController.signOut().catch(console.error)
+              throw error
+            }
+          }
+        } else {
+          await WalletConnectProvider.connect()
+        }
+
         await this.setWalletConnectProvider()
       },
 
@@ -203,19 +248,7 @@ export class Web3Modal extends Web3ModalScaffold {
         info: Info
         provider: Provider
       }) => {
-        if (id === ConstantsUtil.INJECTED_CONNECTOR_ID) {
-          const InjectedProvider = ethersConfig.injected
-          if (!InjectedProvider) {
-            throw new Error('connectionControllerClient:connectInjected - provider is undefined')
-          }
-          try {
-            EthersStoreUtil.setError(undefined)
-            await InjectedProvider.request({ method: 'eth_requestAccounts' })
-            this.setInjectedProvider(ethersConfig)
-          } catch (error) {
-            EthersStoreUtil.setError(error)
-          }
-        } else if (id === ConstantsUtil.EIP6963_CONNECTOR_ID && info && provider) {
+        if (id === ConstantsUtil.EIP6963_CONNECTOR_ID && info && provider) {
           try {
             EthersStoreUtil.setError(undefined)
             await provider.request({ method: 'eth_requestAccounts' })
@@ -246,12 +279,6 @@ export class Web3Modal extends Web3ModalScaffold {
           return Boolean(window.ethereum)
         }
 
-        if (ethersConfig.injected) {
-          if (!window?.ethereum) {
-            return false
-          }
-        }
-
         return ids.some(id => Boolean(window.ethereum?.[String(id)]))
       },
 
@@ -260,15 +287,21 @@ export class Web3Modal extends Web3ModalScaffold {
         const providerType = EthersStoreUtil.state.providerType
         localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
         EthersStoreUtil.reset()
+        if (siweConfig?.options?.signOutOnDisconnect) {
+          const { SIWEController } = await import('@web3modal/siwe')
+          await SIWEController.signOut()
+        }
         if (providerType === ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID) {
           const WalletConnectProvider = provider
           await (WalletConnectProvider as unknown as EthereumProvider).disconnect()
           // eslint-disable-next-line no-negated-condition
-        } else if (providerType !== ConstantsUtil.EMAIL_CONNECTOR_ID) {
-          provider?.emit('disconnect')
-        } else {
+        } else if (providerType === ConstantsUtil.EMAIL_CONNECTOR_ID) {
           await this.emailProvider?.disconnect()
+        } else {
+          provider?.emit('disconnect')
         }
+        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
+        EthersStoreUtil.reset()
       },
 
       signMessage: async (message: string) => {
@@ -290,9 +323,7 @@ export class Web3Modal extends Web3ModalScaffold {
       formatUnits: (value: bigint, decimals: number) => formatUnits(value, decimals),
 
       async estimateGas(data) {
-        const chainId = EthersStoreUtil.state.chainId
-        const provider = EthersStoreUtil.state.provider
-        const address = EthersStoreUtil.state.address
+        const { chainId, provider, address } = EthersStoreUtil.state
 
         if (!provider) {
           throw new Error('connectionControllerClient:sendTransaction - provider is undefined')
@@ -317,16 +348,14 @@ export class Web3Modal extends Web3ModalScaffold {
       },
 
       sendTransaction: async (data: SendTransactionArgs) => {
-        const chainId = EthersStoreUtil.state.chainId
-        const provider = EthersStoreUtil.state.provider
-        const address = EthersStoreUtil.state.address
+        const { chainId, provider, address } = EthersStoreUtil.state
 
         if (!provider) {
-          throw new Error('connectionControllerClient:sendTransaction - provider is undefined')
+          throw new Error('ethersClient:sendTransaction - provider is undefined')
         }
 
         if (!address) {
-          throw new Error('connectionControllerClient:sendTransaction - address is undefined')
+          throw new Error('ethersClient:sendTransaction - address is undefined')
         }
 
         const txParams = {
@@ -344,6 +373,68 @@ export class Web3Modal extends Web3ModalScaffold {
         const txReceipt = await txResponse.wait()
 
         return (txReceipt?.hash as `0x${string}`) || null
+      },
+
+      writeContract: async (data: WriteContractArgs) => {
+        const { chainId, provider, address } = EthersStoreUtil.state
+
+        if (!provider) {
+          throw new Error('ethersClient:writeContract - provider is undefined')
+        }
+
+        if (!address) {
+          throw new Error('ethersClient:writeContract - address is undefined')
+        }
+
+        const browserProvider = new BrowserProvider(provider, chainId)
+        const signer = new JsonRpcSigner(browserProvider, address)
+        const contract = new Contract(data.tokenAddress, data.abi, signer)
+
+        if (!contract || !data.method) {
+          throw new Error('Contract method is undefined')
+        }
+
+        const method = contract[data.method]
+        if (method) {
+          const tx = await method(data.receiverAddress, data.tokenAmount)
+
+          return tx
+        }
+
+        throw new Error('Contract method is undefined')
+      },
+
+      getEnsAddress: async (value: string) => {
+        const { chainId } = EthersStoreUtil.state
+        if (chainId && chainId === 1) {
+          const ensProvider = new InfuraProvider('mainnet')
+
+          const name = await ensProvider.resolveName(value)
+          if (name) {
+            return name
+          }
+
+          return false
+        }
+
+        return false
+      },
+
+      getEnsAvatar: async (value: string) => {
+        const { chainId } = EthersStoreUtil.state
+        if (chainId && chainId === 1) {
+          const ensProvider = new InfuraProvider('mainnet')
+
+          const avatar = await ensProvider.getAvatar(value)
+
+          if (avatar) {
+            return avatar
+          }
+
+          return false
+        }
+
+        return false
       }
     }
 
@@ -410,9 +501,6 @@ export class Web3Modal extends Web3ModalScaffold {
       this.syncEmailConnector(w3mOptions.projectId)
     }
 
-    if (ethersConfig.injected) {
-      this.checkActiveInjectedProvider(ethersConfig)
-    }
     if (ethersConfig.coinbase) {
       this.checkActiveCoinbaseProvider(ethersConfig)
     }
@@ -481,7 +569,7 @@ export class Web3Modal extends Web3ModalScaffold {
     localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
     EthersStoreUtil.reset()
 
-    if (providerType === 'injected' || providerType === 'eip6963') {
+    if (providerType === 'eip6963') {
       provider?.emit('disconnect')
     } else {
       const walletConnectProvider = provider as unknown as EthereumProvider
@@ -570,18 +658,6 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
-  private checkActiveInjectedProvider(config: ProviderType) {
-    const InjectedProvider = config.injected
-    const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
-
-    if (InjectedProvider) {
-      if (walletId === ConstantsUtil.INJECTED_CONNECTOR_ID) {
-        this.setInjectedProvider(config)
-        this.watchInjected(config)
-      }
-    }
-  }
-
   private checkActiveCoinbaseProvider(config: ProviderType) {
     const CoinbaseProvider = config.coinbase as unknown as ExternalProvider
     const walletId = localStorage.getItem(EthersConstantsUtil.WALLET_ID)
@@ -631,23 +707,6 @@ export class Web3Modal extends Web3ModalScaffold {
     }
   }
 
-  private async setInjectedProvider(config: ProviderType) {
-    window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.INJECTED_CONNECTOR_ID)
-    const InjectedProvider = config.injected
-
-    if (InjectedProvider) {
-      const { address, chainId } = await EthersHelpersUtil.getUserInfo(InjectedProvider)
-      if (address && chainId) {
-        EthersStoreUtil.setChainId(chainId)
-        EthersStoreUtil.setProviderType('injected')
-        EthersStoreUtil.setProvider(config.injected)
-        EthersStoreUtil.setIsConnected(true)
-        this.setAddress(address)
-        this.watchCoinbase(config)
-      }
-    }
-  }
-
   private async setEIP6963Provider(provider: Provider, name: string) {
     window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, name)
 
@@ -685,6 +744,7 @@ export class Web3Modal extends Web3ModalScaffold {
     window?.localStorage.setItem(EthersConstantsUtil.WALLET_ID, ConstantsUtil.EMAIL_CONNECTOR_ID)
 
     if (this.emailProvider) {
+      console.log('connecting with emailProvider, chain', this.getChainId())
       const { address, chainId, smartAccountDeployed, preferredAccountType, accounts } =
         await this.emailProvider.connect({ chainId: this.getChainId() })
 
@@ -731,45 +791,6 @@ export class Web3Modal extends Web3ModalScaffold {
     const accountsChangedHandler = async (accounts: string[]) => {
       if (accounts.length > 0) {
         await this.setWalletConnectProvider()
-      }
-    }
-
-    if (provider) {
-      provider.on('disconnect', disconnectHandler)
-      provider.on('accountsChanged', accountsChangedHandler)
-      provider.on('chainChanged', chainChangedHandler)
-    }
-  }
-
-  private watchInjected(config: ProviderType) {
-    const provider = config.injected
-
-    function disconnectHandler() {
-      localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
-      EthersStoreUtil.reset()
-
-      provider?.removeListener('disconnect', disconnectHandler)
-      provider?.removeListener('accountsChanged', accountsChangedHandler)
-      provider?.removeListener('chainChanged', chainChangedHandler)
-    }
-
-    function accountsChangedHandler(accounts: string[]) {
-      const currentAccount = accounts?.[0]
-      if (currentAccount) {
-        EthersStoreUtil.setAddress(getOriginalAddress(currentAccount) as Address)
-      } else {
-        localStorage.removeItem(EthersConstantsUtil.WALLET_ID)
-        EthersStoreUtil.reset()
-      }
-    }
-
-    function chainChangedHandler(chainId: string) {
-      if (chainId) {
-        const chain =
-          typeof chainId === 'string'
-            ? EthersHelpersUtil.hexStringToNumber(chainId)
-            : Number(chainId)
-        EthersStoreUtil.setChainId(chain)
       }
     }
 
@@ -951,6 +972,7 @@ export class Web3Modal extends Web3ModalScaffold {
       const caipAddress: CaipAddress = `${ConstantsUtil.EIP155}:${chainId}:${address}`
 
       this.setIsConnected(isConnected)
+      console.log('@ethers syncAccount', caipAddress)
       this.setPreferredAccountType(preferredAccountType)
       this.setCaipAddress(caipAddress)
       this.syncConnectedWalletInfo()
@@ -1121,29 +1143,6 @@ export class Web3Modal extends Web3ModalScaffold {
             }
           }
         }
-      } else if (providerType === ConstantsUtil.INJECTED_CONNECTOR_ID && chain) {
-        const InjectedProvider = provider
-        if (InjectedProvider) {
-          try {
-            await InjectedProvider.request({
-              method: 'wallet_switchEthereumChain',
-              params: [{ chainId: EthersHelpersUtil.numberToHexString(chain.chainId) }]
-            })
-            EthersStoreUtil.setChainId(chain.chainId)
-            // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          } catch (switchError: any) {
-            if (
-              switchError.code === EthersConstantsUtil.ERROR_CODE_UNRECOGNIZED_CHAIN_ID ||
-              switchError.code === EthersConstantsUtil.ERROR_CODE_DEFAULT ||
-              switchError?.data?.originalError?.code ===
-                EthersConstantsUtil.ERROR_CODE_UNRECOGNIZED_CHAIN_ID
-            ) {
-              await EthersHelpersUtil.addEthereumChain(InjectedProvider, chain)
-            } else {
-              throw new Error('Chain is not supported')
-            }
-          }
-        }
       } else if (providerType === ConstantsUtil.EIP6963_CONNECTOR_ID && chain) {
         const EIP6963Provider = provider
 
@@ -1225,21 +1224,6 @@ export class Web3Modal extends Web3ModalScaffold {
         name: PresetsUtil.ConnectorNamesMap[ConstantsUtil.WALLET_CONNECT_CONNECTOR_ID],
         type: connectorType
       })
-    }
-
-    if (config.injected) {
-      const injectedConnectorType =
-        PresetsUtil.ConnectorTypesMap[ConstantsUtil.INJECTED_CONNECTOR_ID]
-      if (injectedConnectorType) {
-        w3mConnectors.push({
-          id: ConstantsUtil.INJECTED_CONNECTOR_ID,
-          explorerId: PresetsUtil.ConnectorExplorerIds[ConstantsUtil.INJECTED_CONNECTOR_ID],
-          imageId: PresetsUtil.ConnectorImageIds[ConstantsUtil.INJECTED_CONNECTOR_ID],
-          imageUrl: this.options?.connectorImages?.[ConstantsUtil.INJECTED_CONNECTOR_ID],
-          name: PresetsUtil.ConnectorNamesMap[ConstantsUtil.INJECTED_CONNECTOR_ID],
-          type: injectedConnectorType
-        })
-      }
     }
 
     if (config.coinbase) {