+++ packages/ui/src/utils/JSXTypeUtil.ts
@@ -51,6 +51,7 @@ import type { WuiTag } from '../composites/wui-tag/index.js'
 import type { WuiTokenButton } from '../composites/wui-token-button/index.js'
 import type { WuiTokenListItem } from '../composites/wui-token-list-item/index.js'
 import type { WuiTooltip } from '../composites/wui-tooltip/index.js'
+import type { WuiIconButton } from '../composites/wui-icon-button/index.js'
 import type { WuiTransactionListItem } from '../composites/wui-transaction-list-item/index.js'
 import type { WuiTransactionListItemLoader } from '../composites/wui-transaction-list-item-loader/index.js'
 import type { WuiTransactionVisual } from '../composites/wui-transaction-visual/index.js'
@@ -61,7 +62,6 @@ import type { WuiListWalletTransaction } from '../composites/wui-list-wallet-tra
 import type { WuiOnRampActivityItem } from '../composites/wui-onramp-activity-item/index.js'
 import type { WuiOnRampProviderItem } from '../composites/wui-onramp-provider-item/index.js'
 import type { WuiPromo } from '../composites/wui-promo/index.js'
-import type { WuiTooltipSelect } from '../composites/wui-tooltip-select/index.js'
 import type { WuiProfileButton } from '../composites/wui-profile-button/index.js'
 import type { WuiListAccount } from '../composites/wui-list-account/index.js'
 import type { WuiPreviewItem } from '../composites/wui-preview-item/index.js'
@@ -134,7 +134,7 @@ declare global {
       'wui-tag': CustomElement<WuiTag>
       'wui-token-button': CustomElement<WuiTokenButton>
       'wui-token-list-item': CustomElement<WuiTokenListItem>
-      'wui-tooltip-select': CustomElement<WuiTooltipSelect>
+      'wui-icon-button': CustomElement<WuiIconButton>
       'wui-tooltip': CustomElement<WuiTooltip>
       'wui-transaction-list-item-loader': CustomElement<WuiTransactionListItemLoader>
       'wui-transaction-list-item': CustomElement<WuiTransactionListItem>